parisc: Fix panic on OpenBSD/6.7 regarding STI console
OpenBSD complained with a panic ("uvm_mapent_addr_insert: insert
collision with entry...") when starting the STI console.
This happens because of a typo in the STI ROM section entries:
-                *(.sti.textend)
+                *(.sti.text.end)

While at it, move the sti.text.end after the STI fonts and align the STI
region list and built-in font to 32 bytes.

There was a second error in the STI which shows up with wrongly
rendered fonts: The last_char entry should be 255 instead of 0.

Signed-off-by: Helge Deller <deller@gmx.de>