parisc: Alloc any font size with STI font_unpmv function
The font_unpmv function was limited to 8x16 fonts.
This change now allows it to render fonts of any size up to 32x32
pixels.

Millicode functions which do not reside inside the STI code will not get
copied at runtime into userspace RAM, and as such we can not call
millicode functions from within STI code.
But the compiler will usually emit a call to the mulI millicode routine
for any multiplication with an variable integer value, so I had to write
a mulI implementation using shifts, which seems to be pretty good
optimized by the compiler.

Signed-off-by: Helge Deller <deller@gmx.de>