biosvar.h: Add accessors for various BIOS data areas
This patch adds or adjust the existing accessors to the various data
areas in a x86 BIOS for usage with a PA-RISC BIOS.

On PA-RISC the x86 areas are emulated as normal variables, stored in
normal RAM and can be accessed directly.

GET_IVT() and SET_IVT() are emulated by a static ivt_table[] array.

Introduce a get_bda_ptr() inline function to return the address of the
Bios Data Area (BDA).

GET_BDA() and SET_BDA() are used to directly access variables from the
BDA.

GET_GLOBAL() is used to access variables directly.

Signed-off-by: Helge Deller <deller@gmx.de>