blockcmd.c: Prevent unaligned access crash on PA-RISC
The cdbcmd pointer given to scsi_fill_cmd() can point to
an unaligned address. On x86 writing a 64-bit value to an
unaligned address will succeed, while on PA-RISC the machine
will stop with an unaligned access error (esp. since the
fault handlers are not implemented in the firmware).

Work around that issue by using a temporary variable and
copy it to the destination when finished.

Signed-off-by: Helge Deller <deller@gmx.de>