parisc: SEABIOS_HPPA_VERSION 6
New features and fixes:

* Allow "z" instead of "y" in firmware boot menu for german keyboards
* Fixed serial port emulation

Serial ports ttyS0 and ttyS1 were swapped compared to the real hardware.
So, a booting Linux was actually using ttyS1 as serial console, and the
emulated ttyS0 wasn't useable.

With this version you can use e.g.:
    -serial mon:stdio -serial /dev/ttyS4
to use emulated ttyS0 in the gueast for console output, and pass ttyS4
from the host to ttyS1 in the guest.

NOTE: This version is incompatible with older qemu versions, since
the serial ports are now swapped.

Signed-off-by: Helge Deller <deller@gmx.de>