parisc: Fully implement PDC_PSW for 64-bit CPUs
PDC_PSW tells the hardware how the default PSW is implemented.

The w field indicates the default width of the processor.  The w field
also determines whether the External Interrupt Request (EIR) register is
treated as a right-justified 32-bit register or a full 64-bit register.
The e field indicates the default endianness of the processor.  Both
bits determine how the PSW W-bit and E-bit will be set on an
interruption.

Signed-off-by: Helge Deller <deller@gmx.de>